requirements.txt
alembic==1.11.1
ansi2html==1.9.1
appdirs==1.4.4
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.9.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.3
contourpy==1.2.0
cycler==0.12.1
dash==2.14.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
DateTime==5.4
Deprecated==1.2.14
dnspython==2.3.0
email-validator==2.0.0.post2
finance-datareader==0.9.66
Flask==2.3.2
Flask-Caching==2.1.0
Flask-Markdown==0.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.46.0
frozendict==2.3.10
gunicorn==20.1.0
html5lib==1.1
idna==3.4
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lxml==4.9.3
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.8.2
multipledispatch==1.0.0
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.1
pandas==2.1.4
pandas-datareader==0.10.0
peewee==3.17.0
Pillow==10.1.0
plotly==5.15.0
psycopg2-binary==2.9.6
pykrx==1.0.45
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
retrying==1.3.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.17
tabulate==0.9.0
tenacity==8.2.2
tqdm==4.66.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.1.0
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.16.0
WTForms==3.0.1
xlrd==2.0.1
xmltodict==0.13.0
yfinance==0.2.33
zipp==3.17.0
zope.interface==6.1
